[SECURITY] Limit user access in workspace previews
To view a preview of a workspace page a backend user is simulated.
Currently the user who created the preview link is taken into account.
This patch creates a limited backend user to be able to process the
web request.

Besides the security fix, the preview collision of issue #72225 is
resolved with this change as well, see https://review.typo3.org/45253/

Resolves: #28175, #72225
Releases: master, 7.6, 6.2
Change-Id: I7db4827a3fd570ce893449e3f050df68f8734b64