[SECURITY] Prevent insecure TypoScript processing in EXT:form
The form extension uses a TypoScript-like notation to configure mail
forms. Actually this is just pseudo TypoScript that never should be
extended with cObject or stdWrap functionalities, but due to the
similarities these features have been introduced in the past which
might lead to negative side effects.

This change introduces a reduced feature set of ContentObjectRenderer
to still be able to apply common and strict stdWrap functionalities, but
now it is e.g. disallowed to render records from database, call user
functions and process hooks.

Resolves: #69796
Releases: 6.2
Change-Id: Ia199ea04b1492244177fd00eefd7370302a9daf3