[SECURITY] Prevent persistent username in filesystem
The language label for the refresh login popup contains the
username already and is persisted to the filesystem. Use
TYPO3.configuration.username and replace it with JavaScript
instead to prevent the information disclosure.

In 4.5 we can safely remove the sprintf function as the label doesn't
include any placeholder for the user name.

Resolves: #75933
Releases: master, 7.6, 6.2, 4.5
Change-Id: I81277762b42723388c0c628f80fac3dc7a7c9d08